+++ src/Contracts/Output.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlissJaspis\QueryDetector\Contracts;
 
 use Illuminate\Support\Collection;
@@ -7,7 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 
 interface Output
 {
-    public function boot();
+    public function boot(): void;
 
-    public function output(Collection $detectedQueries, Response $response);
+    public function output(Collection $detectedQueries, Response $response): void;
 }