+++ src/Events/QueryDetected.php
@@ -1,18 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlissJaspis\QueryDetector\Events;
 
-use Illuminate\Queue\SerializesModels;
+use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Support\Collection;
 
-class QueryDetected {
-    use SerializesModels;
+class QueryDetected
+{
+    use Dispatchable;
 
     public function __construct(
         public Collection $queries
-    ){}
+    ) {}
 
-    public function getQueries() : Collection
+    public function getQueries(): Collection
     {
         return $this->queries;
     }