+++ src/Outputs/Alert.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlissJaspis\QueryDetector\Outputs;
 
 use BlissJaspis\QueryDetector\Contracts\Output;
@@ -8,12 +10,12 @@ use Symfony\Component\HttpFoundation\Response;
 
 class Alert implements Output
 {
-    public function boot()
+    public function boot(): void
     {
         //
     }
 
-    public function output(Collection $detectedQueries, Response $response)
+    public function output(Collection $detectedQueries, Response $response): void
     {
         if (stripos($response->headers->get('Content-Type'), 'text/html') !== 0 || $response->isRedirection()) {
             return;
@@ -25,26 +27,25 @@ class Alert implements Output
 
         $pos = strripos($content, '</body>');
 
-        if (false !== $pos) {
-            $content = substr($content, 0, $pos) . $outputContent . substr($content, $pos);
+        if ($pos !== false) {
+            $content = substr($content, 0, $pos).$outputContent.substr($content, $pos);
         } else {
-            $content = $content . $outputContent;
+            $content = $content.$outputContent;
         }
 
-        // Update the new content and reset the content length
         $response->setContent($content);
 
         $response->headers->remove('Content-Length');
     }
 
-    protected function getOutputContent(Collection $detectedQueries)
+    protected function getOutputContent(Collection $detectedQueries): string
     {
         $output = '<script type="text/javascript">';
         $output .= "alert('Found the following N+1 queries in this request:\\n\\n";
         foreach ($detectedQueries as $detectedQuery) {
-            $output .= "Model: ".addslashes($detectedQuery['model'])." => Relation: ".addslashes($detectedQuery['relation']);
+            $output .= 'Model: '.addslashes($detectedQuery['model']).' => Relation: '.addslashes($detectedQuery['relation']);
             $output .= " - You should add \"with(\'".addslashes($detectedQuery['relation'])."\')\" to eager-load this relation.";
-            $output .= "\\n";
+            $output .= '\\n';
         }
         $output .= "')";
         $output .= '</script>';