+++ src/Outputs/Clockwork.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlissJaspis\QueryDetector\Outputs;
 
 use BlissJaspis\QueryDetector\Contracts\Output;
@@ -8,12 +10,12 @@ use Symfony\Component\HttpFoundation\Response;
 
 class Clockwork implements Output
 {
-    public function boot()
+    public function boot(): void
     {
         //
     }
 
-    public function output(Collection $detectedQueries, Response $response)
+    public function output(Collection $detectedQueries, Response $response): void
     {
         clock()->warning("{$detectedQueries->count()} N+1 queries detected:", $detectedQueries->toArray());
     }