+++ src/Outputs/Debugbar.php
@@ -1,31 +1,33 @@
 <?php
 
-namespace BlissJaspis\QueryDetector\Outputs;
+declare(strict_types=1);
 
-use Illuminate\Support\Collection;
-use Symfony\Component\HttpFoundation\Response;
+namespace BlissJaspis\QueryDetector\Outputs;
 
 use BlissJaspis\QueryDetector\Contracts\Output;
 use DebugBar\DataCollector\MessagesCollector;
 use Fruitcake\LaravelDebugbar\Facades\Debugbar as LaravelDebugbar;
+use Illuminate\Support\Collection;
+use Symfony\Component\HttpFoundation\Response;
 
 class Debugbar implements Output
 {
-    protected $collector;
+    protected MessagesCollector $collector;
 
-    public function boot()
+    public function boot(): void
     {
         $this->collector = new MessagesCollector('N+1 Queries');
 
-        if (!LaravelDebugbar::hasCollector($this->collector->getName())) {
+        if (! LaravelDebugbar::hasCollector($this->collector->getName())) {
             LaravelDebugbar::addCollector($this->collector);
         }
     }
 
-    public function output(Collection $detectedQueries, Response $response)
+    public function output(Collection $detectedQueries, Response $response): void
     {
         foreach ($detectedQueries as $detectedQuery) {
-            $this->collector->addMessage(sprintf('Model: %s => Relation: %s - You should add `with(%s)` to eager-load this relation.',
+            $this->collector->addMessage(sprintf(
+                'Model: %s => Relation: %s - You should add `with(%s)` to eager-load this relation.',
                 $detectedQuery['model'],
                 $detectedQuery['relation'],
                 $detectedQuery['relation']