+++ workbench/database/migrations/2024_04_30_085322_create_query_detector_tables.php
@@ -1,12 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up() : void
+    public function up(): void
     {
         Schema::create('authors', function (Blueprint $table) {
             $table->bigIncrements('id');
@@ -40,7 +42,7 @@ return new class extends Migration
         });
     }
 
-    public function down() : void
+    public function down(): void
     {
         Schema::dropIfExists('comments');
         Schema::dropIfExists('profiles');