+++ workbench/database/seeders/DatabaseSeeder.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Workbench\Database\Seeders;
 
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 use Workbench\App\Models\Author;
 use Workbench\App\Models\Comment;
@@ -20,11 +21,11 @@ class DatabaseSeeder extends Seeder
 
         $authors->each(function ($author) {
             Profile::factory()->create([
-                'author_id' => $author->id
+                'author_id' => $author->id,
             ]);
 
             $posts = Post::factory()->count(5)->create([
-                'author_id' => $author->id
+                'author_id' => $author->id,
             ]);
 
             $posts->each(function ($post) {